+++ src/App.js
@@ -1,5 +1,5 @@
 import React, { useEffect } from "react";
-import CanvasElement from "./components/canvas";
+import MatchingGame from "./components/matching-game";
 import './App.css';
 
 
@@ -12,16 +12,18 @@ function App() {
     style={{
       width: "100%",
       height: "100vh",
-      backgroundColor: "pink"
+      backgroundColor: "lightgreen"
     }}
     >
       {/*<a href="https://www.flaticon.com/free-icons/maps-and-location" title="maps and location icons">
         Maps and location icons created by Freepik - Flaticon
   </a>*/}
-      <CanvasElement
+      <button></button>
+      <MatchingGame
         width={400} //800
-        height={"100vh"} //800
+        height={800} //800
         color={"pink"}
+        renderMode={"all"}
       />
     </div>
   );